+++ .ansible/collections/ansible_collections/netapp/ontap/CHANGELOG.md
@@ -10,6 +10,7 @@
   - na_ontap_snapshot - new option `snaplock_expiry_time` added in REST, requires ONTAP 9.15.1 or later.
   - na_ontap_volume_clone - new options `time_out`, `wait_for_completion` added in REST.
   - na_ontap_security_ssh - new option `host_key_algorithms`, requires ONTAP 9.16.1 or later.
+  - na_ontap_rest_info - added error handling when API doesn't return zero records.
   - updated `README` template; added 'Support' section.
 
 ### Bug Fixes